requirements.txt
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
exceptiongroup==1.3.1
feedparser==6.0.12
html2text==2025.4.15
idna==3.11
iniconfig==2.1.0
lxml==6.0.2
packaging==26.0
pluggy==1.6.0
pygments==2.19.2
pytest==8.4.2
pyyaml==6.0.3
requests==2.32.5
sgmllib3k==1.0.0
soupsieve==2.8.3
tomli==2.4.0
typing-extensions==4.15.0
urllib3==2.6.3
